bdlim: Bayesian Distributed Lag Interaction Models

Estimation and interpretation of Bayesian distributed lag interaction models (BDLIMs). A BDLIM regresses a scalar outcome on repeated measures of exposure and allows for modification by a categorical variable under four specific patterns of modification. The main function is bdlim(). There are also summary and plotting files. Details on methodology are described in Wilson et al. (2017) doi:10.1093/biostatistics/kxx002 .

Details

The main function is bdlim4.

References

Wilson A, Chiu YM, Hsu HL, Wright RO, Wright RJ, Coull BA (2017). “Bayesian distributed lag interaction models to identify perinatal windows of vulnerability in children's health.” Biostatistics, 18(3), 537–552. doi:10.1093/biostatistics/kxx002 . arXiv preprint.

See also

Author

Maintainer: Ander Wilson ander.wilson@colostate.edu (ORCID)