Get weight function

Get weight function

Usage

getw(object, type = "normal", ...)

Arguments

object

An object of class bdlim1.

type

Type of summary. The default is 'normal' which returns a summary with the mean projected to conform with the norm 1 constraint. Others are raw which is the same as normal but without the mean projected and full that returns all posterior samples used in the analysis (burn in removed and thinned).

...

Not used.

Value

A data frame with either the summary of the weight function (types normal and raw) or a posterior sample of the weight function (type full).