Simulated Birth Data

A dataset containing simulated birth data for examples with bdlim. Add outcome and covariate data is simulated. The exposure data is real exposure data. Therefore, it has realistic correlation structure. The exposures are consistent with the date of conception variables.

Usage

sbd_bdlim

Format

A data frame with 1000 rows (observations) and 202 variables:

bwgaz

Outcome to be used. Simulated birth weight for gestational age z-score.

ChildSex

Binary sex of child.

MomAge

Continuous age in years.

GestAge

Continuous estimated gestational age at birth in weeks.

MomHeightIn

Continuous maternal height in inches.

MomPriorWeightLbs

Continuous mothers pre-pregnancy weight in pounds.

MomPriorBMI

Continuous mothers pre-pregnancy BMI.

race

Categorical race.

Hispanic

Binary indicator of Hispanic.

MomEdu

Categorical maternal heighest educational attainment.

SmkAny

Binary indicator of any smoking during pregnancy.

Marital

Categorical maternal marital status.

Income

Categorical income.

EstDateConcept

Estimated date of conception.

EstMonthConcept

Estimated month of conception.

EstYearConcept

Estimated year of conception.

pm25_1

Exposure to be used. Weekly average exposure to PM2.5 in week 1 of gestation.

pm25_2

Exposure to be used. Weekly average exposure to PM2.5 in week 2 of gestation.

pm25_3

Exposure to be used. Weekly average exposure to PM2.5 in week 3 of gestation.

pm25_4

Exposure to be used. Weekly average exposure to PM2.5 in week 4 of gestation.

pm25_5

Exposure to be used. Weekly average exposure to PM2.5 in week 5 of gestation.

pm25_6

Exposure to be used. Weekly average exposure to PM2.5 in week 6 of gestation.

pm25_7

Exposure to be used. Weekly average exposure to PM2.5 in week 7 of gestation.

pm25_8

Exposure to be used. Weekly average exposure to PM2.5 in week 8 of gestation.

pm25_9

Exposure to be used. Weekly average exposure to PM2.5 in week 9 of gestation.

pm25_10

Exposure to be used. Weekly average exposure to PM2.5 in week 10 of gestation.

pm25_11

Exposure to be used. Weekly average exposure to PM2.5 in week 11 of gestation.

pm25_12

Exposure to be used. Weekly average exposure to PM2.5 in week 12 of gestation.

pm25_13

Exposure to be used. Weekly average exposure to PM2.5 in week 13 of gestation.

pm25_14

Exposure to be used. Weekly average exposure to PM2.5 in week 14 of gestation.

pm25_15

Exposure to be used. Weekly average exposure to PM2.5 in week 15 of gestation.

pm25_16

Exposure to be used. Weekly average exposure to PM2.5 in week 16 of gestation.

pm25_17

Exposure to be used. Weekly average exposure to PM2.5 in week 17 of gestation.

pm25_18

Exposure to be used. Weekly average exposure to PM2.5 in week 18 of gestation.

pm25_19

Exposure to be used. Weekly average exposure to PM2.5 in week 19 of gestation.

pm25_20

Exposure to be used. Weekly average exposure to PM2.5 in week 20 of gestation.

pm25_21

Exposure to be used. Weekly average exposure to PM2.5 in week 21 of gestation.

pm25_22

Exposure to be used. Weekly average exposure to PM2.5 in week 22 of gestation.

pm25_23

Exposure to be used. Weekly average exposure to PM2.5 in week 23 of gestation.

pm25_24

Exposure to be used. Weekly average exposure to PM2.5 in week 24 of gestation.

pm25_25

Exposure to be used. Weekly average exposure to PM2.5 in week 25 of gestation.

pm25_26

Exposure to be used. Weekly average exposure to PM2.5 in week 26 of gestation.

pm25_27

Exposure to be used. Weekly average exposure to PM2.5 in week 27 of gestation.

pm25_28

Exposure to be used. Weekly average exposure to PM2.5 in week 28 of gestation.

pm25_29

Exposure to be used. Weekly average exposure to PM2.5 in week 29 of gestation.

pm25_30

Exposure to be used. Weekly average exposure to PM2.5 in week 30 of gestation.

pm25_31

Exposure to be used. Weekly average exposure to PM2.5 in week 31 of gestation.

pm25_32

Exposure to be used. Weekly average exposure to PM2.5 in week 32 of gestation.

pm25_33

Exposure to be used. Weekly average exposure to PM2.5 in week 33 of gestation.

pm25_34

Exposure to be used. Weekly average exposure to PM2.5 in week 34 of gestation.

pm25_35

Exposure to be used. Weekly average exposure to PM2.5 in week 35 of gestation.

pm25_36

Exposure to be used. Weekly average exposure to PM2.5 in week 36 of gestation.

pm25_37

Exposure to be used. Weekly average exposure to PM2.5 in week 37 of gestation.

no2_1

Exposure to be used. Weekly average exposure to NO2 in week 1 of gestation.

no2_2

Exposure to be used. Weekly average exposure to NO2 in week 2 of gestation.

no2_3

Exposure to be used. Weekly average exposure to NO2 in week 3 of gestation.

no2_4

Exposure to be used. Weekly average exposure to NO2 in week 4 of gestation.

no2_5

Exposure to be used. Weekly average exposure to NO2 in week 5 of gestation.

no2_6

Exposure to be used. Weekly average exposure to NO2 in week 6 of gestation.

no2_7

Exposure to be used. Weekly average exposure to NO2 in week 7 of gestation.

no2_8

Exposure to be used. Weekly average exposure to NO2 in week 8 of gestation.

no2_9

Exposure to be used. Weekly average exposure to NO2 in week 9 of gestation.

no2_10

Exposure to be used. Weekly average exposure to NO2 in week 10 of gestation.

no2_11

Exposure to be used. Weekly average exposure to NO2 in week 11 of gestation.

no2_12

Exposure to be used. Weekly average exposure to NO2 in week 12 of gestation.

no2_13

Exposure to be used. Weekly average exposure to NO2 in week 13 of gestation.

no2_14

Exposure to be used. Weekly average exposure to NO2 in week 14 of gestation.

no2_15

Exposure to be used. Weekly average exposure to NO2 in week 15 of gestation.

no2_16

Exposure to be used. Weekly average exposure to NO2 in week 16 of gestation.

no2_17

Exposure to be used. Weekly average exposure to NO2 in week 17 of gestation.

no2_18

Exposure to be used. Weekly average exposure to NO2 in week 18 of gestation.

no2_19

Exposure to be used. Weekly average exposure to NO2 in week 19 of gestation.

no2_20

Exposure to be used. Weekly average exposure to NO2 in week 20 of gestation.

no2_21

Exposure to be used. Weekly average exposure to NO2 in week 21 of gestation.

no2_22

Exposure to be used. Weekly average exposure to NO2 in week 22 of gestation.

no2_23

Exposure to be used. Weekly average exposure to NO2 in week 23 of gestation.

no2_24

Exposure to be used. Weekly average exposure to NO2 in week 24 of gestation.

no2_25

Exposure to be used. Weekly average exposure to NO2 in week 25 of gestation.

no2_26

Exposure to be used. Weekly average exposure to NO2 in week 26 of gestation.

no2_27

Exposure to be used. Weekly average exposure to NO2 in week 27 of gestation.

no2_28

Exposure to be used. Weekly average exposure to NO2 in week 28 of gestation.

no2_29

Exposure to be used. Weekly average exposure to NO2 in week 29 of gestation.

no2_30

Exposure to be used. Weekly average exposure to NO2 in week 30 of gestation.

no2_31

Exposure to be used. Weekly average exposure to NO2 in week 31 of gestation.

no2_32

Exposure to be used. Weekly average exposure to NO2 in week 32 of gestation.

no2_33

Exposure to be used. Weekly average exposure to NO2 in week 33 of gestation.

no2_34

Exposure to be used. Weekly average exposure to NO2 in week 34 of gestation.

no2_35

Exposure to be used. Weekly average exposure to NO2 in week 35 of gestation.

no2_36

Exposure to be used. Weekly average exposure to NO2 in week 36 of gestation.

no2_37

Exposure to be used. Weekly average exposure to NO2 in week 37 of gestation.

so2_1

Exposure to be used. Weekly average exposure to SO2 in week 1 of gestation.

so2_2

Exposure to be used. Weekly average exposure to SO2 in week 2 of gestation.

so2_3

Exposure to be used. Weekly average exposure to SO2 in week 3 of gestation.

so2_4

Exposure to be used. Weekly average exposure to SO2 in week 4 of gestation.

so2_5

Exposure to be used. Weekly average exposure to SO2 in week 5 of gestation.

so2_6

Exposure to be used. Weekly average exposure to SO2 in week 6 of gestation.

so2_7

Exposure to be used. Weekly average exposure to SO2 in week 7 of gestation.

so2_8

Exposure to be used. Weekly average exposure to SO2 in week 8 of gestation.

so2_9

Exposure to be used. Weekly average exposure to SO2 in week 9 of gestation.

so2_10

Exposure to be used. Weekly average exposure to SO2 in week 10 of gestation.

so2_11

Exposure to be used. Weekly average exposure to SO2 in week 11 of gestation.

so2_12

Exposure to be used. Weekly average exposure to SO2 in week 12 of gestation.

so2_13

Exposure to be used. Weekly average exposure to SO2 in week 13 of gestation.

so2_14

Exposure to be used. Weekly average exposure to SO2 in week 14 of gestation.

so2_15

Exposure to be used. Weekly average exposure to SO2 in week 15 of gestation.

so2_16

Exposure to be used. Weekly average exposure to SO2 in week 16 of gestation.

so2_17

Exposure to be used. Weekly average exposure to SO2 in week 17 of gestation.

so2_18

Exposure to be used. Weekly average exposure to SO2 in week 18 of gestation.

so2_19

Exposure to be used. Weekly average exposure to SO2 in week 19 of gestation.

so2_20

Exposure to be used. Weekly average exposure to SO2 in week 20 of gestation.

so2_21

Exposure to be used. Weekly average exposure to SO2 in week 21 of gestation.

so2_22

Exposure to be used. Weekly average exposure to SO2 in week 22 of gestation.

so2_23

Exposure to be used. Weekly average exposure to SO2 in week 23 of gestation.

so2_24

Exposure to be used. Weekly average exposure to SO2 in week 24 of gestation.

so2_25

Exposure to be used. Weekly average exposure to SO2 in week 25 of gestation.

so2_26

Exposure to be used. Weekly average exposure to SO2 in week 26 of gestation.

so2_27

Exposure to be used. Weekly average exposure to SO2 in week 27 of gestation.

so2_28

Exposure to be used. Weekly average exposure to SO2 in week 28 of gestation.

so2_29

Exposure to be used. Weekly average exposure to SO2 in week 29 of gestation.

so2_30

Exposure to be used. Weekly average exposure to SO2 in week 30 of gestation.

so2_31

Exposure to be used. Weekly average exposure to SO2 in week 31 of gestation.

so2_32

Exposure to be used. Weekly average exposure to SO2 in week 32 of gestation.

so2_33

Exposure to be used. Weekly average exposure to SO2 in week 33 of gestation.

so2_34

Exposure to be used. Weekly average exposure to SO2 in week 34 of gestation.

so2_35

Exposure to be used. Weekly average exposure to SO2 in week 35 of gestation.

so2_36

Exposure to be used. Weekly average exposure to SO2 in week 36 of gestation.

so2_37

Exposure to be used. Weekly average exposure to SO2 in week 37 of gestation.

co_1

Exposure to be used. Weekly average exposure to CO in week 1 of gestation.

co_2

Exposure to be used. Weekly average exposure to CO in week 2 of gestation.

co_3

Exposure to be used. Weekly average exposure to CO in week 3 of gestation.

co_4

Exposure to be used. Weekly average exposure to CO in week 4 of gestation.

co_5

Exposure to be used. Weekly average exposure to CO in week 5 of gestation.

co_6

Exposure to be used. Weekly average exposure to CO in week 6 of gestation.

co_7

Exposure to be used. Weekly average exposure to CO in week 7 of gestation.

co_8

Exposure to be used. Weekly average exposure to CO in week 8 of gestation.

co_9

Exposure to be used. Weekly average exposure to CO in week 9 of gestation.

co_10

Exposure to be used. Weekly average exposure to CO in week 10 of gestation.

co_11

Exposure to be used. Weekly average exposure to CO in week 11 of gestation.

co_12

Exposure to be used. Weekly average exposure to CO in week 12 of gestation.

co_13

Exposure to be used. Weekly average exposure to CO in week 13 of gestation.

co_14

Exposure to be used. Weekly average exposure to CO in week 14 of gestation.

co_15

Exposure to be used. Weekly average exposure to CO in week 15 of gestation.

co_16

Exposure to be used. Weekly average exposure to CO in week 16 of gestation.

co_17

Exposure to be used. Weekly average exposure to CO in week 17 of gestation.

co_18

Exposure to be used. Weekly average exposure to CO in week 18 of gestation.

co_19

Exposure to be used. Weekly average exposure to CO in week 19 of gestation.

co_20

Exposure to be used. Weekly average exposure to CO in week 20 of gestation.

co_21

Exposure to be used. Weekly average exposure to CO in week 21 of gestation.

co_22

Exposure to be used. Weekly average exposure to CO in week 22 of gestation.

co_23

Exposure to be used. Weekly average exposure to CO in week 23 of gestation.

co_24

Exposure to be used. Weekly average exposure to CO in week 24 of gestation.

co_25

Exposure to be used. Weekly average exposure to CO in week 25 of gestation.

co_26

Exposure to be used. Weekly average exposure to CO in week 26 of gestation.

co_27

Exposure to be used. Weekly average exposure to CO in week 27 of gestation.

co_28

Exposure to be used. Weekly average exposure to CO in week 28 of gestation.

co_29

Exposure to be used. Weekly average exposure to CO in week 29 of gestation.

co_30

Exposure to be used. Weekly average exposure to CO in week 30 of gestation.

co_31

Exposure to be used. Weekly average exposure to CO in week 31 of gestation.

co_32

Exposure to be used. Weekly average exposure to CO in week 32 of gestation.

co_33

Exposure to be used. Weekly average exposure to CO in week 33 of gestation.

co_34

Exposure to be used. Weekly average exposure to CO in week 34 of gestation.

co_35

Exposure to be used. Weekly average exposure to CO in week 35 of gestation.

co_36

Exposure to be used. Weekly average exposure to CO in week 36 of gestation.

co_37

Exposure to be used. Weekly average exposure to CO in week 37 of gestation.

temp_1

Exposure to be used. Weekly average exposure to temperature in week 1 of gestation.

temp_2

Exposure to be used. Weekly average exposure to temperature in week 2 of gestation.

temp_3

Exposure to be used. Weekly average exposure to temperature in week 3 of gestation.

temp_4

Exposure to be used. Weekly average exposure to temperature in week 4 of gestation.

temp_5

Exposure to be used. Weekly average exposure to temperature in week 5 of gestation.

temp_6

Exposure to be used. Weekly average exposure to temperature in week 6 of gestation.

temp_7

Exposure to be used. Weekly average exposure to temperature in week 7 of gestation.

temp_8

Exposure to be used. Weekly average exposure to temperature in week 8 of gestation.

temp_9

Exposure to be used. Weekly average exposure to temperature in week 9 of gestation.

temp_10

Exposure to be used. Weekly average exposure to temperature in week 10 of gestation.

temp_11

Exposure to be used. Weekly average exposure to temperature in week 11 of gestation.

temp_12

Exposure to be used. Weekly average exposure to temperature in week 12 of gestation.

temp_13

Exposure to be used. Weekly average exposure to temperature in week 13 of gestation.

temp_14

Exposure to be used. Weekly average exposure to temperature in week 14 of gestation.

temp_15

Exposure to be used. Weekly average exposure to temperature in week 15 of gestation.

temp_16

Exposure to be used. Weekly average exposure to temperature in week 16 of gestation.

temp_17

Exposure to be used. Weekly average exposure to temperature in week 17 of gestation.

temp_18

Exposure to be used. Weekly average exposure to temperature in week 18 of gestation.

temp_19

Exposure to be used. Weekly average exposure to temperature in week 19 of gestation.

temp_20

Exposure to be used. Weekly average exposure to temperature in week 20 of gestation.

temp_21

Exposure to be used. Weekly average exposure to temperature in week 21 of gestation.

temp_22

Exposure to be used. Weekly average exposure to temperature in week 22 of gestation.

temp_23

Exposure to be used. Weekly average exposure to temperature in week 23 of gestation.

temp_24

Exposure to be used. Weekly average exposure to temperature in week 24 of gestation.

temp_25

Exposure to be used. Weekly average exposure to temperature in week 25 of gestation.

temp_26

Exposure to be used. Weekly average exposure to temperature in week 26 of gestation.

temp_27

Exposure to be used. Weekly average exposure to temperature in week 27 of gestation.

temp_28

Exposure to be used. Weekly average exposure to temperature in week 28 of gestation.

temp_29

Exposure to be used. Weekly average exposure to temperature in week 29 of gestation.

temp_30

Exposure to be used. Weekly average exposure to temperature in week 30 of gestation.

temp_31

Exposure to be used. Weekly average exposure to temperature in week 31 of gestation.

temp_32

Exposure to be used. Weekly average exposure to temperature in week 32 of gestation.

temp_33

Exposure to be used. Weekly average exposure to temperature in week 33 of gestation.

temp_34

Exposure to be used. Weekly average exposure to temperature in week 34 of gestation.

temp_35

Exposure to be used. Weekly average exposure to temperature in week 35 of gestation.

temp_36

Exposure to be used. Weekly average exposure to temperature in week 36 of gestation.

temp_37

Exposure to be used. Weekly average exposure to temperature in week 37 of gestation.

source

Variable indicating that the data came from the bdlim package.