Simulated Birth Data

A dataset containing simulated birth data for examples with bdlim. Add outcome and covariate data is simulated. The exposure data is real exposure data. Therefore, it has realistic correlation structure. The exposures are consistent with the date of conception variables.

Usage

simulated_birth_data_c1

Format

A data frame with 1000 rows (observations) and 51 variables:

bwgaz

Outcome to be used. Simulated birth weight for gestational age z-score.

ChildSex

Binary sex of child.

MomAge

Continuous age in years.

GestAge

Continuous estimated gestational age at birth in weeks.

MomHeightIn

Continuous maternal height in inches.

MomPriorWeightLbs

Continuous mothers pre-pregnancy weight in pounds.

MomPriorBMI

Continuous mothers pre-pregnancy BMI.

race

Categorical race.

Hispanic

Binary indicator of Hispanic.

MomEdu

Categorical maternal heighest educational attainment.

SmkAny

Binary indicator of any smoking during pregnancy.

Marital

Categorical maternal marital status.

Income

Categorical income.

EstDateConcept

Estimated date of conception.

EstMonthConcept

Estimated month of conception.

EstYearConcept

Estimated year of conception.

pm25_1

Exposure to be used. Weekly average exposure to PM2.5 in week 1 of gestation.

pm25_2

Exposure to be used. Weekly average exposure to PM2.5 in week 2 of gestation.

pm25_3

Exposure to be used. Weekly average exposure to PM2.5 in week 3 of gestation.

pm25_4

Exposure to be used. Weekly average exposure to PM2.5 in week 4 of gestation.

pm25_5

Exposure to be used. Weekly average exposure to PM2.5 in week 5 of gestation.

pm25_6

Exposure to be used. Weekly average exposure to PM2.5 in week 6 of gestation.

pm25_7

Exposure to be used. Weekly average exposure to PM2.5 in week 7 of gestation.

pm25_8

Exposure to be used. Weekly average exposure to PM2.5 in week 8 of gestation.

pm25_9

Exposure to be used. Weekly average exposure to PM2.5 in week 9 of gestation.

pm25_10

Exposure to be used. Weekly average exposure to PM2.5 in week 10 of gestation.

pm25_11

Exposure to be used. Weekly average exposure to PM2.5 in week 11 of gestation.

pm25_12

Exposure to be used. Weekly average exposure to PM2.5 in week 12 of gestation.

pm25_13

Exposure to be used. Weekly average exposure to PM2.5 in week 13 of gestation.

pm25_14

Exposure to be used. Weekly average exposure to PM2.5 in week 14 of gestation.

pm25_15

Exposure to be used. Weekly average exposure to PM2.5 in week 15 of gestation.

pm25_16

Exposure to be used. Weekly average exposure to PM2.5 in week 16 of gestation.

pm25_17

Exposure to be used. Weekly average exposure to PM2.5 in week 17 of gestation.

pm25_18

Exposure to be used. Weekly average exposure to PM2.5 in week 18 of gestation.

pm25_19

Exposure to be used. Weekly average exposure to PM2.5 in week 19 of gestation.

pm25_20

Exposure to be used. Weekly average exposure to PM2.5 in week 20 of gestation.

pm25_21

Exposure to be used. Weekly average exposure to PM2.5 in week 21 of gestation.

pm25_22

Exposure to be used. Weekly average exposure to PM2.5 in week 22 of gestation.

pm25_23

Exposure to be used. Weekly average exposure to PM2.5 in week 23 of gestation.

pm25_24

Exposure to be used. Weekly average exposure to PM2.5 in week 24 of gestation.

pm25_25

Exposure to be used. Weekly average exposure to PM2.5 in week 25 of gestation.

pm25_26

Exposure to be used. Weekly average exposure to PM2.5 in week 26 of gestation.

pm25_27

Exposure to be used. Weekly average exposure to PM2.5 in week 27 of gestation.

pm25_28

Exposure to be used. Weekly average exposure to PM2.5 in week 28 of gestation.

pm25_29

Exposure to be used. Weekly average exposure to PM2.5 in week 29 of gestation.

pm25_30

Exposure to be used. Weekly average exposure to PM2.5 in week 30 of gestation.

pm25_31

Exposure to be used. Weekly average exposure to PM2.5 in week 31 of gestation.

pm25_32

Exposure to be used. Weekly average exposure to PM2.5 in week 32 of gestation.

pm25_33

Exposure to be used. Weekly average exposure to PM2.5 in week 33 of gestation.

pm25_34

Exposure to be used. Weekly average exposure to PM2.5 in week 34 of gestation.

pm25_35

Exposure to be used. Weekly average exposure to PM2.5 in week 35 of gestation.

pm25_36

Exposure to be used. Weekly average exposure to PM2.5 in week 36 of gestation.

pm25_37

Exposure to be used. Weekly average exposure to PM2.5 in week 37 of gestation.