Summary for bdlim4

Summary for bdlim4

Usage

# S3 method for bdlim4
summary(object, model = NULL, ...)

Arguments

object

An object of class bdlim4.

model

Pattern of heterogeneity to be printed. If not specified (default) the best fitting model will be used. Options are "n", "b", "w" and "bw" where b indicates the effect sizes are subgroup specific and w indicates the weight functions are subgroups specific.

...

Other arguments

Value

An object of class summary.bdlim2.

Examples

# run BDLIM with modification by ChildSex
fit_sex <- bdlim4(
  y = sbd_bdlim$bwgaz,
  exposure = sbd_bdlim[,paste0("pm25_",1:37)],
  covars = sbd_bdlim[,c("MomPriorBMI","MomAge","race","Hispanic",
                                      "EstMonthConcept","EstYearConcept")],
  group = as.factor(sbd_bdlim$ChildSex),
  df = 5,
  nits = 5000,
  parallel = FALSE
)
#> fitting bw
#> fitting b
#> fitting w
#> fitting n
#> postprocessingpostprocessing

#summarize results
summary(fit_sex)
#> 
#> Call:
#> bdlim4(y = sbd_bdlim$bwgaz, exposure = sbd_bdlim[, paste0("pm25_", 
#>     1:37)], covars = sbd_bdlim[, c("MomPriorBMI", "MomAge", "race", 
#>     "Hispanic", "EstMonthConcept", "EstYearConcept")], group = as.factor(sbd_bdlim$ChildSex), 
#>     df = 5, nits = 5000, parallel = FALSE)
#> 
#> 
#> Model fit statistics:
#>   *bw*      b      w      n 
#> 0.5296 0.3628 0.1076 0.0000 
#> 
#> 
#> Estimated cumulative effects:
#>  group        mean      median         sd       q2.5       q97.5 pr_gr0
#>      F -0.02432105 -0.01440156 0.05840875 -0.1589347  0.08582361 0.3668
#>      M -0.41422629 -0.41269390 0.07649831 -0.5668490 -0.26614927 0.0000
#> 
#> 
#> Estimated covariate regression coefficients:
#>                 name         mean       median          sd         q2.5
#>           interceptF  5.434554698  5.311275643 6.207368319 -6.574748240
#>           interceptM  5.662387763  5.513271742 6.216510984 -6.287204843
#>          MomPriorBMI -0.015793585 -0.015780369 0.003257048 -0.022233043
#>               MomAge  0.001106069  0.001187026 0.003037259 -0.004954562
#>          raceAsianPI -0.024981039 -0.023775522 0.173210082 -0.377356368
#>            raceBlack -0.106174285 -0.105140143 0.175510245 -0.450673891
#>            racewhite -0.053981212 -0.051906842 0.167421971 -0.394601789
#>  HispanicNonHispanic  0.254235274  0.254601670 0.038832853  0.176360545
#>     EstMonthConcept2 -0.153929864 -0.154078335 0.095824770 -0.337651347
#>     EstMonthConcept3 -0.098920397 -0.098827761 0.096841426 -0.289975140
#>     EstMonthConcept4 -0.158305414 -0.161166068 0.092152807 -0.330664997
#>     EstMonthConcept5 -0.090417089 -0.091420176 0.086527424 -0.261534802
#>     EstMonthConcept6 -0.188681080 -0.185999727 0.083583588 -0.349027862
#>     EstMonthConcept7 -0.043542514 -0.043718668 0.088412873 -0.217936886
#>     EstMonthConcept8  0.163815698  0.167005406 0.099256189 -0.034993787
#>     EstMonthConcept9  0.337285607  0.336741936 0.093743041  0.153899097
#>    EstMonthConcept10  0.442123165  0.442935080 0.095511214  0.254795869
#>    EstMonthConcept11  0.241005241  0.241286077 0.087693897  0.065900278
#>    EstMonthConcept12  0.081323860  0.081218695 0.082969410 -0.079213931
#>       EstYearConcept -0.001987777 -0.001948197 0.003080323 -0.008231377
#>         q97.5 pr_gr0
#>  18.004813152 0.8108
#>  18.407301078 0.8212
#>  -0.009277323 0.0000
#>   0.006934105 0.6472
#>   0.301639790 0.4476
#>   0.227100569 0.2704
#>   0.264974838 0.3708
#>   0.329948558 1.0000
#>   0.032204697 0.0588
#>   0.094468144 0.1500
#>   0.022003348 0.0488
#>   0.080495546 0.1504
#>  -0.026090260 0.0120
#>   0.126264129 0.3136
#>   0.359503991 0.9440
#>   0.520593258 1.0000
#>   0.630994032 1.0000
#>   0.417777532 0.9964
#>   0.244136722 0.8332
#>   0.003908005 0.2548
#> 
#> 
#> BDLIM fit on 1000 observations. Estimated residual standard deviation is 0.542 (0.517,0.566). WAIC is 1634.925.
#> 
#> Use `plot(); for the summary.bdlim4 object to view estimated distributed lag functions. The `dlfun' object in the summary object contains estimates of the lag functions.

# obtain estimates of the distributed lag function
# these are note displayed when printed but available for use
sfit_sex <- summary(fit_sex)
head(sfit_sex$dlfun)
#>   group time          mean        median          sd        q2.5       q97.5
#> 1     F    1  0.0002848911  2.201839e-04 0.008708333 -0.01905842 0.018619913
#> 2     F    2 -0.0006538565 -6.548615e-05 0.006978144 -0.01667180 0.013407866
#> 3     F    3 -0.0015652764 -5.852380e-04 0.005530421 -0.01486478 0.008823013
#> 4     F    4 -0.0024220410 -1.363822e-03 0.004572943 -0.01362232 0.005266576
#> 5     F    5 -0.0031968228 -2.258604e-03 0.004279074 -0.01351167 0.003458295
#> 6     F    6 -0.0038622940 -2.915773e-03 0.004561568 -0.01459347 0.002787865
#>   pr_gr0
#> 1 0.5304
#> 2 0.4908
#> 3 0.4196
#> 4 0.3072
#> 5 0.2208
#> 6 0.1920

# can summarize with a specific model
sfit_hisp_n <- summary(fit_sex, model="n") # no modification
sfit_hisp_b <- summary(fit_sex, model="b") # subgroup-specific effects (beta)
sfit_hisp_w <- summary(fit_sex, model="w") # subgroup-specific weight function
sfit_hisp_bw <- summary(fit_sex, model="bw") # both subgroup-specific